+++ malachite-base/src/bin_util/demo_and_bench/strings/latex.rs
@@ -6,12 +6,17 @@
 // Lesser General Public License (LGPL) as published by the Free Software Foundation; either version
 // 3 of the License, or (at your option) any later version. See <https://www.gnu.org/licenses/>.
 
+use malachite_base::strings::latex::ToLatex;
+use malachite_base::test_util::generators::common::{GenConfig, GenMode};
+use malachite_base::test_util::generators::string_gen;
 use malachite_base::test_util::runner::Runner;
 
 pub(crate) fn register(runner: &mut Runner) {
-    crement::register(runner);
-    latex::register(runner);
+    register_demo!(runner, demo_str_to_latex);
 }
 
-mod crement;
-mod latex;
+fn demo_str_to_latex(gm: GenMode, config: &GenConfig, limit: usize) {
+    for s in string_gen().get(gm, config).take(limit) {
+        println!("{:?}.to_latex() = {}", s, s.to_latex());
+    }
+}