+++ malachite-base/src/bin_util/demo_and_bench/strings/mod.rs
@@ -10,12 +10,14 @@ use malachite_base::test_util::runner::Runner;
 
 pub(crate) fn register(runner: &mut Runner) {
     gmp_format::register(runner);
+    latex::register(runner);
     string_is_subset::register(runner);
     string_sort::register(runner);
     string_unique::register(runner);
 }
 
 mod gmp_format;
+mod latex;
 mod string_is_subset;
 mod string_sort;
 mod string_unique;